+++ drift_dev/lib/src/analyzer/dart/column_parser.dart
@@ -29,12 +29,14 @@ const String _methodDefault = 'withDefault';
 const String _methodClientDefault = 'clientDefault';
 const String _methodMap = 'map';
 const String _methodGenerated = 'generatedAs';
+const String _methodCheck = 'check';
 const Set<String> _addsSqlConstraint = {
   _methodReferences,
   _methodAutoIncrement,
   _methodUnique,
   _methodDefault,
-  _methodGenerated
+  _methodGenerated,
+  _methodCheck,
 };
 
 const String _errorMessage = 'This getter does not create a valid column that '
@@ -326,6 +328,10 @@ class ColumnParser {
                 .substring(generatedExpression.offset, generatedExpression.end);
             generatedAs = ColumnGeneratedAs(code, stored);
           }
+          break;
+        case _methodCheck:
+          final expr = remainingExpr.argumentList.arguments.first.toSource();
+          foundFeatures.add(DartCheckExpression(expr));
       }
 
       // We're not at a starting method yet, so we need to go deeper!